% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs-suite.R
\name{cs_get_suite}
\alias{cs_get_suite}
\title{Get a suite definition}
\usage{
cs_get_suite(suite_id)
}
\arguments{
\item{suite_id}{Character scalar identifying the suite.}
}
\value{
A character vector of DGP IDs contained in the suite.
}
\description{
Get a suite definition
}
