% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs-constants.R
\docType{data}
\name{cs_tau_oracle}
\alias{cs_tau_oracle}
\title{Oracle quantile grid}
\format{
An object of class \code{numeric} of length 99.
}
\usage{
cs_tau_oracle
}
\value{
Numeric vector of quantile levels from 0.01 to 0.99.
}
\description{
Constant sequence of tau values used for oracle quantile targets.
}
\keyword{datasets}
