% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs-accessors.R
\name{cs_tidy_batch}
\alias{cs_tidy_batch}
\title{Tidy a batch of run results}
\usage{
cs_tidy_batch(batch_results)
}
\arguments{
\item{batch_results}{A list of run result objects (from \code{cs_read_batch()}).}
}
\value{
A tibble of tidy results.
}
\description{
Applies \code{cs_tidy()} to each result object in a batch and returns a single
combined tibble.
}
