% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs-staging.R
\name{cs_gather_results}
\alias{cs_gather_results}
\title{Gather staged results and pin them to a board}
\usage{
cs_gather_results(board, staging_dir)
}
\arguments{
\item{board}{A pins board.}

\item{staging_dir}{Directory containing staged .qs files.}
}
\value{
Number of files gathered and pinned.
}
\description{
Gather staged results and pin them to a board
}
