% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est-ipw-att.R
\name{est_ipw_att}
\alias{est_ipw_att}
\title{Inverse-probability weighted ATT estimator (IPW-ATT)}
\usage{
est_ipw_att(df, config = list(), tau = cs_tau_oracle, ...)
}
\description{
Inverse-probability weighted ATT estimator (IPW-ATT)
}
