% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est-lm-att.R
\name{est_lm_att}
\alias{est_lm_att}
\title{Linear outcome regression ATT estimator}
\usage{
est_lm_att(df, tau = cs_tau_oracle, config = list())
}
\description{
Fits a linear model on controls and uses g-computation to estimate ATT.
}
