% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs-helpers.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing helper}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to test.}

\item{y}{Fallback if \code{x} is \code{NULL}.}
}
\value{
\code{x} or \code{y}.
}
\description{
Returns \code{x} if it is not \code{NULL}, otherwise returns \code{y}.
}
